refactor(api): move OrganizationInvitedUser model to src/team
