+++ api/lib/domain/models/index.js
@@ -89,7 +89,6 @@ import { KnowledgeElement } from './KnowledgeElement.js';
 import { LearningContent } from './LearningContent.js';
 import { AreaForAdmin, CompetenceForAdmin, ThematicForAdmin, TubeForAdmin } from './LearningContentForAdmin.js';
 import { Membership } from './Membership.js';
-import { OrganizationInvitedUser } from './OrganizationInvitedUser.js';
 import { OrganizationLearner } from './OrganizationLearner.js';
 import { OrganizationMemberIdentity } from './OrganizationMemberIdentity.js';
 import { OrganizationTag } from './OrganizationTag.js';
@@ -204,7 +203,6 @@ export {
   Organization,
   OrganizationForAdmin,
   OrganizationInvitation,
-  OrganizationInvitedUser,
   OrganizationLearner,
   OrganizationMemberIdentity,
   OrganizationPlacesLot,
