+++ api/tests/team/unit/domain/usecases/accept-organization-invitation.usecase.test.js
@@ -1,7 +1,7 @@
-import { AlreadyExistingMembershipError } from '../../../../../src/shared/domain/errors.js';
-import { OrganizationInvitedUser } from '../../../../../src/team/domain/models/OrganizationInvitedUser.js';
-import { acceptOrganizationInvitation } from '../../../../../src/team/domain/usecases/accept-organization-invitation.usecase.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
+import { OrganizationInvitedUser } from '../../../../src/team/domain/models/OrganizationInvitedUser.js';
+import { acceptOrganizationInvitation } from '../../../../src/team/domain/usecases/accept-organization-invitation.usecase.js';
+import { AlreadyExistingMembershipError } from '../../../../src/shared/domain/errors.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | Domain | UseCases | accept-organization-invitation', function () {
   let organizationInvitedUserRepository;
